Fix: Implemented Rate Games and Review Games
1. Members create reviews from game detail page in pending status
2. Member gets reviews only owned by themselves
3. SuperAdmin/Admin/Moderator get all reviews
4. SuperAdmin/Admin/Moderator updates the review to processed status
5. UI change to font awesome stars rating in Create/Edit/Delete Review

+++ PROG3050/Controllers/ReviewsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -15,17 +16,31 @@ namespace PROG3050.Controllers
     public class ReviewsController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<User> _userManager;
 
-        public ReviewsController(ApplicationDbContext context)
+        public ReviewsController(ApplicationDbContext context, UserManager<User> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Reviews
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Review.Include(r => r.Game).Include(r => r.User);
-            return View(await applicationDbContext.ToListAsync());
+            var currentUser = await _userManager.GetUserAsync(User);
+            var roles = await _userManager.GetRolesAsync(currentUser);
+            if (roles.Contains("SuperAdmin") || roles.Contains("Admin") || roles.Contains("Moderator"))
+            {
+                var reviews = _context.Review.Include(r => r.Game).Include(r => r.User);
+                return View(await reviews.ToListAsync());
+            }
+            else
+            {
+                var reviews = _context.Review.Include(r => r.Game)
+                                            .Include(r => r.User)
+                                            .Where(r => r.UserId == currentUser.Id);
+                return View(await reviews.ToListAsync());
+            }
         }
 
         // GET: Reviews/Details/5
@@ -45,20 +60,32 @@ namespace PROG3050.Controllers
                 return NotFound();
             }
 
-            return View(review);
+            // Validate the review owned by the current user or SuperAdmin/Admin/Moderator
+            var currentUser = await _userManager.GetUserAsync(User);
+            if (review.UserId == currentUser.Id || User.IsInRole("SuperAdmin") || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+            {
+                return View(review);
+            }
+
+            TempData["GlobalStatusMessage"] = $"You don't have permission to the review.";
+            TempData["GlobalStatusMessageClass"] = "danger";
+            return RedirectToAction(nameof(Index));
         }
 
+        [Authorize(Roles = "SuperAdmin,Admin,Moderator")]
         // GET: Reviews/Create
         public IActionResult Create()
         {
             ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title");
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName");
+            ViewData["Status"] = new SelectList(new List<string>() { "Pending", "Processed" });
             return View();
         }
 
         // POST: Reviews/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize(Roles = "SuperAdmin,Admin,Moderator")]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ReviewId,Title,Description,Rating,Status,UserId,GameId")] Review review)
@@ -70,7 +97,8 @@ namespace PROG3050.Controllers
                 return RedirectToAction(nameof(Index));
             }
             ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title", review.GameId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", review.UserId);
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", review.UserId);
+            ViewData["Status"] = new SelectList(new List<string>() { "Pending", "Processed" }, review.Status);
             return View(review);
         }
 
@@ -83,9 +111,13 @@ namespace PROG3050.Controllers
             {
                 _context.Add(review);
                 await _context.SaveChangesAsync();
+                
+                TempData["GlobalStatusMessage"] = $"Your review has been requested sucessfully. Admin will take a look and process it shortly.";
+                TempData["GlobalStatusMessageClass"] = "success";
+                
                 return RedirectToAction("Details", "Games", new { id = review.GameId });
             }
-            TempData["GlobalStatusMessage"] = $"Title and Description are required.";
+            TempData["GlobalStatusMessage"] = $"Title, Description, and Rating are required.";
             TempData["GlobalStatusMessageClass"] = "danger";
             return RedirectToAction("Details", "Games", new { id = review.GameId });
         }
@@ -103,9 +135,20 @@ namespace PROG3050.Controllers
             {
                 return NotFound();
             }
-            ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title", review.GameId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", review.UserId);
-            return View(review);
+
+            // Validate the review owned by the current user or SuperAdmin/Admin/Moderator
+            var currentUser = await _userManager.GetUserAsync(User);
+            if (review.UserId == currentUser.Id || User.IsInRole("SuperAdmin") || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+            {
+                ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title", review.GameId);
+                ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", review.UserId);
+                ViewData["Status"] = new SelectList(new List<string>() { "Pending", "Processed" }, review.Status);
+                return View(review);
+            }
+
+            TempData["GlobalStatusMessage"] = $"You don't have permission to the review.";
+            TempData["GlobalStatusMessageClass"] = "danger";
+            return RedirectToAction(nameof(Index));            
         }
 
         // POST: Reviews/Edit/5
@@ -120,29 +163,39 @@ namespace PROG3050.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            // Validate the review owned by the current user or SuperAdmin/Admin/Moderator
+            var currentUser = await _userManager.GetUserAsync(User);
+            if (review.UserId == currentUser.Id || User.IsInRole("SuperAdmin") || User.IsInRole("Admin") || User.IsInRole("Moderator"))
             {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(review);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ReviewExists(review.ReviewId))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(review);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!ReviewExists(review.ReviewId))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title", review.GameId);
+                ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", review.UserId);
+                ViewData["Status"] = new SelectList(new List<string>() { "Pending", "Processed" }, review.Status);
+                return View(review);
             }
-            ViewData["GameId"] = new SelectList(_context.Game, "GameId", "Title", review.GameId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", review.UserId);
-            return View(review);
+
+            TempData["GlobalStatusMessage"] = $"You don't have permission to the review.";
+            TempData["GlobalStatusMessageClass"] = "danger";
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Reviews/Delete/5
@@ -162,7 +215,16 @@ namespace PROG3050.Controllers
                 return NotFound();
             }
 
-            return View(review);
+            // Validate the review owned by the current user or SuperAdmin/Admin/Moderator
+            var currentUser = await _userManager.GetUserAsync(User);
+            if (review.UserId == currentUser.Id || User.IsInRole("SuperAdmin") || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+            {
+                return View(review);
+            }
+
+            TempData["GlobalStatusMessage"] = $"You don't have permission to the review.";
+            TempData["GlobalStatusMessageClass"] = "danger";
+            return RedirectToAction(nameof(Index));
         }
 
         // POST: Reviews/Delete/5
@@ -175,9 +237,20 @@ namespace PROG3050.Controllers
                 return Problem("Entity set 'ApplicationDbContext.Review'  is null.");
             }
             var review = await _context.Review.FindAsync(id);
+
             if (review != null)
             {
-                _context.Review.Remove(review);
+                // Validate the review owned by the current user or SuperAdmin/Admin/Moderator
+                var currentUser = await _userManager.GetUserAsync(User);
+                if (review.UserId == currentUser.Id || User.IsInRole("SuperAdmin") || User.IsInRole("Admin") || User.IsInRole("Moderator"))
+                {
+                    _context.Review.Remove(review);
+                }
+                else
+                {
+                    TempData["GlobalStatusMessage"] = $"You don't have permission to the review.";
+                    TempData["GlobalStatusMessageClass"] = "danger";
+                }
             }
             
             await _context.SaveChangesAsync();
